refactor: split dto package to v1/v2
