+++ src/main/java/site/archive/api/v1/ArchiveControllerV1.java
@@ -14,10 +14,10 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
-import site.archive.api.dto.archive.ArchiveCountResponseDto;
-import site.archive.api.dto.archive.ArchiveDto;
-import site.archive.api.dto.archive.ArchiveImageUrlResponseDto;
-import site.archive.api.dto.archive.ArchiveListResponseDto;
+import site.archive.api.v1.dto.archive.ArchiveCountResponseDto;
+import site.archive.api.v1.dto.archive.ArchiveDto;
+import site.archive.api.v1.dto.archive.ArchiveImageUrlResponseDto;
+import site.archive.api.v1.dto.archive.ArchiveListResponseDto;
 import site.archive.api.resolver.annotation.RequestUser;
 import site.archive.domain.archive.ArchiveImageService;
 import site.archive.domain.archive.ArchiveService;
