+++ src/main/java/site/archive/api/v1/RegisterControllerV1.java
@@ -11,7 +11,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import site.archive.api.command.PasswordRegisterCommand;
-import site.archive.api.dto.user.OAuthRegisterRequestDto;
+import site.archive.api.v1.dto.user.OAuthRegisterRequestDto;
 import site.archive.domain.user.UserRegisterService;
 import site.archive.domain.user.info.UserInfo;
 import site.archive.infra.user.oauth.OAuthUserService;
