+++ src/main/java/site/archive/api/v1/UserControllerV1.java
@@ -12,9 +12,9 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import site.archive.api.command.LoginCommand;
-import site.archive.api.dto.archive.EmailDuplicateResponseDto;
-import site.archive.api.dto.user.UserEmailRequestDto;
-import site.archive.api.dto.user.UserPasswordResetRequestDto;
+import site.archive.api.v1.dto.archive.EmailDuplicateResponseDto;
+import site.archive.api.v1.dto.user.UserEmailRequestDto;
+import site.archive.api.v1.dto.user.UserPasswordResetRequestDto;
 import site.archive.api.resolver.annotation.RequestUser;
 import site.archive.domain.user.UserAuthService;
 import site.archive.domain.user.UserService;
