+++ src/main/java/site/archive/api/v1/dto/archive/ArchiveImageUrlResponseDto.java
@@ -1,4 +1,4 @@
-package site.archive.api.dto.archive;
+package site.archive.api.v1.dto.archive;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
