+++ src/main/java/site/archive/api/v1/dto/user/UserPasswordResetRequestDto.java
@@ -1,4 +1,4 @@
-package site.archive.api.dto.user;
+package site.archive.api.v1.dto.user;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
