+++ src/main/java/site/archive/api/v2/ArchiveCommunityControllerV2.java
@@ -7,7 +7,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import site.archive.api.dto.archive.ArchiveCommunityResponseDto;
+import site.archive.api.v2.dto.ArchiveCommunityResponseDto;
 import site.archive.domain.archive.ArchiveCommunityService;
 import site.archive.domain.archive.ArchiveCommunityTimeSortType;
 import site.archive.domain.archive.entity.Emotion;
