+++ src/main/java/site/archive/api/v2/ArchiveControllerV2.java
@@ -8,8 +8,8 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import site.archive.api.dto.archive.ArchiveDto;
-import site.archive.api.dto.archive.ArchiveListResponseDto;
+import site.archive.api.v1.dto.archive.ArchiveDto;
+import site.archive.api.v1.dto.archive.ArchiveListResponseDto;
 import site.archive.api.resolver.annotation.RequestUser;
 import site.archive.domain.archive.ArchiveService;
 import site.archive.domain.user.info.UserInfo;
