+++ src/main/java/site/archive/domain/archive/ArchiveService.java
@@ -3,8 +3,8 @@ package site.archive.domain.archive;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import site.archive.api.dto.archive.ArchiveDto;
-import site.archive.api.dto.archive.ArchiveListResponseDto;
+import site.archive.api.v1.dto.archive.ArchiveDto;
+import site.archive.api.v1.dto.archive.ArchiveListResponseDto;
 import site.archive.domain.archive.entity.Archive;
 import site.archive.domain.user.UserRepository;
 import site.archive.domain.user.info.UserInfo;
