+++ src/main/java/site/archive/domain/common/MessagingService.java
@@ -1,6 +1,6 @@
 package site.archive.domain.common;
 
-import site.archive.api.dto.user.BaseUserDto;
+import site.archive.api.v1.dto.user.BaseUserDto;
 
 public interface MessagingService {
