+++ src/main/java/site/archive/domain/user/UserRegisterService.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import site.archive.api.command.BasicRegisterCommand;
 import site.archive.api.command.OAuthRegisterCommand;
-import site.archive.api.dto.user.BaseUserDto;
+import site.archive.api.v1.dto.user.BaseUserDto;
 import site.archive.domain.common.MessagingService;
 import site.archive.domain.user.entity.BaseUser;
 import site.archive.domain.user.info.UserInfo;
