+++ src/main/java/site/archive/infra/user/oauth/OAuthUserService.java
@@ -4,7 +4,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.security.authentication.ProviderNotFoundException;
 import org.springframework.stereotype.Service;
 import site.archive.api.command.OAuthRegisterCommand;
-import site.archive.api.dto.user.OAuthRegisterRequestDto;
+import site.archive.api.v1.dto.user.OAuthRegisterRequestDto;
 import site.archive.infra.user.oauth.provider.OAuthProviderClient;
 
 import java.util.List;
