+++ src/main/java/site/archive/infra/user/oauth/provider/OAuthProviderClient.java
@@ -1,7 +1,7 @@
 package site.archive.infra.user.oauth.provider;
 
 import site.archive.api.command.OAuthRegisterCommand;
-import site.archive.api.dto.user.OAuthRegisterRequestDto;
+import site.archive.api.v1.dto.user.OAuthRegisterRequestDto;
 
 public interface OAuthProviderClient {
