+++ src/main/java/site/archive/infra/user/oauth/provider/dto/KakaoProviderRequirements.java
@@ -2,7 +2,7 @@ package site.archive.infra.user.oauth.provider.dto;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
-import site.archive.api.dto.user.OAuthRegisterRequestDto;
+import site.archive.api.v1.dto.user.OAuthRegisterRequestDto;
 
 @RequiredArgsConstructor
 @Getter
