+++ src/test/java/site/archive/infra/user/oauth/provider/KakaoClientTest.java
@@ -7,7 +7,7 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.client.RestTemplate;
-import site.archive.api.dto.user.OAuthRegisterRequestDto;
+import site.archive.api.v1.dto.user.OAuthRegisterRequestDto;
 import site.archive.domain.user.entity.OAuthProvider;
 import site.archive.exception.user.OAuthRegisterFailException;
 import site.archive.infra.user.oauth.provider.dto.KakaoUserInfo;
